refactor: move cmp
