+++ src/frontend/src/flows/addDevice/manage/addDeviceSuccess.json
